+++ internal/app/apptainer/plugin_compile_linux.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2018-2021, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -20,7 +20,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"runtime"
-	"runtime/debug"
 	"strings"
 
 	"github.com/apptainer/apptainer/internal/pkg/buildcfg"
@@ -28,11 +27,11 @@ import (
 	"github.com/apptainer/apptainer/internal/pkg/util/bin"
 	pluginapi "github.com/apptainer/apptainer/pkg/plugin"
 	"github.com/apptainer/apptainer/pkg/sylog"
-	"github.com/apptainer/apptainer/pkg/util/archive"
 	"github.com/apptainer/sif/v2/pkg/sif"
 )
 
-const version = "v0.0.0"
+// source file that should be present in a valid Apptainer source tree
+const canaryFile = "pkg/plugin/plugin.go"
 
 const goVersionFile = `package main
 import "fmt"
@@ -40,37 +39,16 @@ import "runtime"
 func main() { fmt.Printf(runtime.Version()) }`
 
 type buildToolchain struct {
-	goPath          string
-	apptainerSource string
-	pluginDir       string
-	buildTags       string
-	envs            []string
-}
-
-func getPackageName() string {
-	if buildInfo, ok := debug.ReadBuildInfo(); ok {
-		return buildInfo.Main.Path
-	}
-	return "github.com/apptainer/apptainer"
+	workPath  string
+	goPath    string
+	buildTags string
+	envs      []string
 }
 
 // getApptainerSrcDir returns the source directory for apptainer.
 func getApptainerSrcDir() (string, error) {
 	dir := buildcfg.SOURCEDIR
-	pkgName := getPackageName()
-
-	// get current file path
-	_, filename, _, ok := runtime.Caller(0)
-	if !ok {
-		return "", fmt.Errorf("could not determine source directory")
-	}
-
-	// replace github.com/apptainer/apptainer@v0.0.0
-	pattern := fmt.Sprintf("%s@%s", pkgName, version)
-	filename = strings.Replace(filename, pattern, "", 1)
-
-	// look if source directory is present
-	canary := filepath.Join(dir, filename)
+	canary := filepath.Join(dir, canaryFile)
 	sylog.Debugf("Searching source file %s", canary)
 
 	switch _, err := os.Stat(canary); {
@@ -88,8 +66,14 @@ func getApptainerSrcDir() (string, error) {
 // checkGoVersion returns an error if the currently Go toolchain is
 // different from the one used to compile apptainer. Apptainer
 // and plugin must be compiled with the same toolchain.
-func checkGoVersion(tmpDir, goPath string) error {
+func checkGoVersion(goPath string) error {
 	var out bytes.Buffer
+
+	tmpDir, err := ioutil.TempDir("", "plugin-")
+	if err != nil {
+		return errors.New("temporary directory creation failed")
+	}
+	defer os.RemoveAll(tmpDir)
 
 	path := filepath.Join(tmpDir, "rt_version.go")
 	if err := ioutil.WriteFile(path, []byte(goVersionFile), 0o600); err != nil {
@@ -130,94 +114,63 @@ func pluginManifestPath(sourceDir string) string {
 // CompilePlugin compiles a plugin. It takes as input: sourceDir, the path to the
 // plugin's source code directory; and destSif, the path to the intended final
 // location of the plugin SIF file.
-func CompilePlugin(sourceDir, destSif, buildTags string, disableMinorCheck bool) error {
-	apptainerSrcDir, err := getApptainerSrcDir()
+func CompilePlugin(sourceDir, destSif, buildTags string) error {
+	apptainerSrc, err := getApptainerSrcDir()
 	if err != nil {
-		return errors.New("apptainer source directory not found")
+		return fmt.Errorf("apptainer source directory not usable: %w", err)
 	}
+	apptainerSrc, err = filepath.Abs(apptainerSrc)
+	if err != nil {
+		return fmt.Errorf("while getting absolute path of %q: %w", apptainerSrc, err)
+	}
+	sylog.Debugf("Using apptainer source: %s", apptainerSrc)
+	pluginSrc, err := filepath.Abs(sourceDir)
+	if err != nil {
+		return fmt.Errorf("while getting absolute path of %q: %w", pluginSrc, err)
+	}
+	sylog.Debugf("Using plugin source: %s", pluginSrc)
+	if !strings.HasPrefix(pluginSrc, apptainerSrc+string(os.PathSeparator)) {
+		return fmt.Errorf("plugin source %q must be inside apptainer source %q", pluginSrc, apptainerSrc)
+	}
+
 	goPath, err := bin.FindBin("go")
 	if err != nil {
 		return errors.New("go compiler not found")
 	}
 
-	// copy plugin directory to apply modification on-the-fly
-	d, err := ioutil.TempDir("", "plugin-")
-	if err != nil {
-		return errors.New("temporary directory creation failed")
-	}
-	defer os.RemoveAll(d)
-
 	// we need to use the exact same go runtime version used
 	// to compile Apptainer
-	if err := checkGoVersion(d, goPath); err != nil {
+	if err := checkGoVersion(goPath); err != nil {
 		return fmt.Errorf("while checking go version: %s", err)
 	}
 
-	pluginDir := filepath.Join(d, "src")
-
-	err = archive.CopyWithTar(sourceDir, pluginDir)
-	if err != nil {
-		return fmt.Errorf("while copying with tar: %v", err)
-	}
-
-	sourceLink := filepath.Join(pluginDir, plugin.ApptainerSource)
-	// delete it first if already present
-	sylog.Debugf("removing symlink: %s", sourceLink)
-	if err := os.Remove(sourceLink); err != nil {
-		sylog.Warningf("removing symlink: error %s", err.Error())
-		if !errors.Is(err, fs.ErrNotExist) {
-			return fmt.Errorf("while removing symlink named %s: %v", sourceLink, err)
-		}
-	}
-
-	sylog.Warningf("adding new symlink: %s", sourceLink)
-	if err := os.Symlink(apptainerSrcDir, sourceLink); err != nil {
-		sylog.Warningf("adding new symlink: error %s", err.Error())
-		return fmt.Errorf("while creating %s symlink: %s", sourceLink, err)
-	}
-
 	bTool := buildToolchain{
-		buildTags:       buildTags,
-		apptainerSource: apptainerSrcDir,
-		pluginDir:       pluginDir,
-		goPath:          goPath,
-		envs:            append(os.Environ(), "GO111MODULE=on"),
-	}
-
-	// generating final go.mod file
-	modData, err := plugin.PrepareGoModules(sourceDir, disableMinorCheck)
-	if err != nil {
-		return err
-	}
-
-	goMod := filepath.Join(pluginDir, "go.mod")
-	if err := ioutil.WriteFile(goMod, modData, 0o600); err != nil {
-		return fmt.Errorf("while generating %s: %s", goMod, err)
-	}
-
-	// running go mod tidy for plugin go.sum and cleanup
-	var e bytes.Buffer
-	cmd := exec.Command(goPath, "mod", "tidy")
-	cmd.Stderr = &e
-	cmd.Dir = pluginDir
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("while verifying module: %s\nCommand error:\n%s", err, e.String())
+		buildTags: buildTags,
+		workPath:  apptainerSrc,
+		goPath:    goPath,
+		envs:      append(os.Environ(), "GO111MODULE=on"),
 	}
 
 	// build plugin object using go build
-	if _, err := buildPlugin(pluginDir, bTool); err != nil {
+	soPath, err := buildPlugin(pluginSrc, bTool)
+	if err != nil {
 		return fmt.Errorf("while building plugin .so: %v", err)
 	}
+	defer os.Remove(soPath)
 
 	// generate plugin manifest from .so
-	if err := generateManifest(pluginDir, bTool); err != nil {
+	mPath, err := generateManifest(pluginSrc, bTool)
+	if err != nil {
 		return fmt.Errorf("while generating plugin manifest: %s", err)
 	}
+	defer os.Remove(mPath)
 
 	// convert the built plugin object into a sif
-	if err := makeSIF(pluginDir, destSif); err != nil {
+	if err := makeSIF(pluginSrc, destSif); err != nil {
 		return fmt.Errorf("while making sif file: %s", err)
 	}
+
+	sylog.Infof("Plugin built to: %s", destSif)
 
 	return nil
 }
@@ -243,14 +196,14 @@ func buildPlugin(sourceDir string, bTool buildToolchain) (string, error) {
 		"-trimpath",
 		"-buildmode=plugin",
 		"-tags", bTool.buildTags,
-		".",
+		sourceDir,
 	}
 
 	sylog.Debugf("Running: %s %s", bTool.goPath, strings.Join(args, " "))
 
 	buildcmd := exec.Command(bTool.goPath, args...)
 
-	buildcmd.Dir = bTool.pluginDir
+	buildcmd.Dir = bTool.workPath
 	buildcmd.Stderr = os.Stderr
 	buildcmd.Stdout = os.Stdout
 	buildcmd.Stdin = os.Stdin
@@ -262,26 +215,26 @@ func buildPlugin(sourceDir string, bTool buildToolchain) (string, error) {
 // generateManifest takes the path to the plugin source, extracts
 // plugin's manifest by loading it into memory and stores it's json
 // representation in a separate file.
-func generateManifest(sourceDir string, bTool buildToolchain) error {
+func generateManifest(sourceDir string, bTool buildToolchain) (string, error) {
 	in := pluginObjPath(sourceDir)
 	out := pluginManifestPath(sourceDir)
 
 	p, err := plugin.LoadObject(in)
 	if err != nil {
-		return fmt.Errorf("while loading plugin %s: %s", in, err)
+		return "", fmt.Errorf("while loading plugin %s: %s", in, err)
 	}
 
 	f, err := os.OpenFile(out, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0o644)
 	if err != nil {
-		return fmt.Errorf("while creating manifest %s: %s", out, err)
+		return "", fmt.Errorf("while creating manifest %s: %s", out, err)
 	}
 	defer f.Close()
 
 	if err := json.NewEncoder(f).Encode(p.Manifest); err != nil {
-		return fmt.Errorf("while writing manifest %s: %s", out, err)
+		return "", fmt.Errorf("while writing manifest %s: %s", out, err)
 	}
 
-	return nil
+	return out, nil
 }
 
 // makeSIF takes in two arguments: sourceDir, the path to the plugin source directory;